requirements.txt
numpy
pandas
scipy
QuantLib-Python
matplotlib
jupyterlab 